evaluation/fid/requirements.txt
numpy
scipy
pillow
tqdm
safe-gpu
requests
tensorflow[and-cuda]>=2.0
